Project Complete

+++ app/src/main/java/com/example/weatherapi/model/Main.kt
@@ -3,7 +3,7 @@ package com.example.weatherapi.model
 import java.io.Serializable
 
 data class Main(
-    val temp: Int,
+    val temp: Float,
     val pressure: Int,
     val humidity: Int,
     val tempMin: Int,
